PDF_Q&A_project/requirements.txt
PyPDF2
google-generativeai
langchain
langchain-google-genai
streamlit
python-dotenv
langchain-community
faiss-cpu
